Set Visibility for Second Stage of Form
I have setup a visibility checker so that the second stage of the form remains hidden until the user has filled out the first 3 sections. I have used a state to control what is shown and an onclick on the first 3 items that calls it to check to see if the state changes.