Thanks Update
I have updated the structure using more components. I have also updated the thanks note. Instead of sending a p tag back through which looses the body I have used a modal component instead.

This required me to download modal using yarn and then apply it with "import Modal from 'react-modal';" This has allowed me to use a basic popup screen that is triggered when the state of form submitted is flipped to true. It can be removed by clicking the button, clicking off the screen, or presing esc.