Added new Express redirector
I have added a new command in ExpressJS that sends the client to the posted forms page when the form is submitted. The details of the client go to the console (Command Promt).

I have also removed Validator from the project as it is not used. React-modal has also been removed as I have used a simple design on another route to be the thank you note.