Cleaned Up Folder
I have removed some false issues and also remove live-server as it won't get used.

I have also added the url to the console when it runs so it is clear where the user has to go to use the app.

By removing live-server I have also had to update the readme file so that the tool is no longer seen.